+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changes
 - `-i/--ignore_input` and `-c/--check` are now mutually exclusive
 - removed titles from mangadex links
+- preserved order of input for `down`, `conf -a/--add-tracked`
+- more verbose output for `conf -r/--remove-tracked` if argument is not tracked
 
 ### Internal
 - all site modules now inherit from BaseManga